refactor: fix linting warnings for fast refresh
