+++ src/contexts/InternalFirebaseContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const InternalFirebaseContext = createContext(null);
