+++ src/contexts/MapContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const MapContext = createContext(null);
