+++ src/contexts/RemoteConfigContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+// exported for use in storybook preview.jsx
+
+export const RemoteConfigContext = createContext(null);
